Extract chat proactive greeting core
