+++ _api/chat.ts
@@ -1,13 +1,11 @@
 import type { VercelRequest, VercelResponse } from "@vercel/node";
 import {
   streamText,
-  generateText,
   smoothStream,
   convertToModelMessages,
   stepCountIs,
   type UIMessage,
 } from "ai";
-import { google } from "@ai-sdk/google";
 import {
   SupportedModel,
   DEFAULT_MODEL,
@@ -22,7 +20,7 @@ import {
   TOOL_USAGE_INSTRUCTIONS,
   MEMORY_INSTRUCTIONS,
 } from "./_utils/_aiPrompts.js";
-import { getMemoryIndex, getDailyNotesForPrompt, getUnprocessedDailyNotesExcludingToday, type MemoryIndex } from "./_utils/_memory.js";
+import { getMemoryIndex, getDailyNotesForPrompt, type MemoryIndex } from "./_utils/_memory.js";
 import { SUPPORTED_AI_MODELS } from "./_utils/_aiModels.js";
 import { checkAndIncrementAIMessageCount } from "./_utils/_rate-limit.js";
 import { validateAuth } from "./_utils/auth/index.js";
@@ -31,6 +29,7 @@ import { initLogger } from "./_utils/_logging.js";
 import { isAllowedOrigin, getEffectiveOrigin, setCorsHeaders } from "./_utils/_cors.js";
 import { getRequestGeolocation } from "./_utils/_geolocation.js";
 import { createChatTools } from "./chat/tools/index.js";
+import { executeChatProactiveGreetingCore } from "./cores/chat-proactive-greeting-core.js";
 
 // Helper to ensure messages are in UIMessage format for AI SDK v6
 // Handles both simple { role, content } format and UIMessage format with parts
@@ -677,99 +676,19 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
     // -------------------------------------------------------------
     if (isProactiveGreeting && username && validationResult.valid) {
       log("Proactive greeting requested");
-
-      // Background: process past daily notes into long-term memory (fire-and-forget)
-      // Only triggered on proactive greetings (once per session) to avoid
-      // redundant checks on every chat message.
-      try {
-        getUnprocessedDailyNotesExcludingToday(redis, username).then(async (unprocessedNotes) => {
-          if (unprocessedNotes.length > 0) {
-            log(`[DailyNotes] Found ${unprocessedNotes.length} unprocessed past daily notes for ${username}, triggering background processing`);
-            const { processDailyNotesForUser } = await import("./ai/process-daily-notes.js");
-            processDailyNotesForUser(redis, username, log, logError).catch((err: unknown) => {
-              logError("[DailyNotes] Background processing failed (non-blocking):", err);
-            });
-          }
-        }).catch(() => {});
-      } catch { /* non-blocking */ }
-
-      // Build memory context
-      let memoryContext = "";
-      if (userMemories && userMemories.memories.length > 0) {
-        memoryContext += "## User's long-term memories:\n";
-        for (const mem of userMemories.memories) {
-          memoryContext += `- ${mem.key}: ${mem.summary}\n`;
-        }
-      }
-      if (dailyNotesText) {
-        memoryContext += `\n## Recent daily notes:\n${dailyNotesText}\n`;
-      }
-
-      // If no memories, return null (client will keep the generic greeting)
-      if (!memoryContext) {
-        log("No memories available for proactive greeting");
-        res.setHeader("Access-Control-Allow-Origin", validOrigin);
-        res.status(200).json({ greeting: null, reason: "no memories available" });
-        return;
-      }
-
-      // Time context
-      const now = new Date();
-      const sfTime = now.toLocaleTimeString("en-US", {
-        timeZone: "America/Los_Angeles",
-        hour: "numeric",
-        minute: "2-digit",
-        hour12: true,
-      });
-      const dayOfWeek = now.toLocaleDateString("en-US", {
-        timeZone: "America/Los_Angeles",
-        weekday: "long",
+      const proactiveResult = await executeChatProactiveGreetingCore({
+        redis,
+        username,
+        isAuthenticated: validationResult.valid,
+        userMemories,
+        dailyNotesText,
+        log,
+        logError,
       });
 
-      try {
-        const { text } = await generateText({
-          model: google("gemini-2.5-flash"),
-          temperature: 1,
-          maxTokens: 150,
-          system: `You are Ryo, a friendly AI assistant. You're greeting a returning user at the start of a new chat.
-
-Your style:
-- Lowercase, casual, warm
-- Short (1-2 sentences max, under 30 words)
-- No emojis unless natural
-- Sound like a close friend checking in, not a corporate assistant
-- Don't be cheesy or over-enthusiastic
-- Be specific — reference something from their memories or recent activity
-- Mix it up: sometimes ask a question, sometimes share an observation, sometimes reference a shared interest
-
-It's ${dayOfWeek} ${sfTime}. The user's name is "${username}".
-
-${memoryContext}
-
-Generate ONE short proactive greeting. Pick one interesting angle from the context — a recent topic, a memory, something timely — and use it naturally. Don't try to cover everything.
-
-Examples of good greetings:
-- "hey, how's the cursor roadmap coming along?"
-- "morning — did you ever try that restaurant you mentioned?"
-- "back again. still working on that project?"
-- "hey ryo. happy friday — any plans?"
-
-Do NOT start with generic greetings like "hey! i'm ryo" or "welcome back". Jump straight into something specific and interesting. Output ONLY the greeting text, nothing else.`,
-          prompt: "Generate a proactive greeting.",
-        });
-
-        const greeting = text.trim();
-        log(`Generated proactive greeting: "${greeting.substring(0, 50)}..."`);
-
-        res.setHeader("Access-Control-Allow-Origin", validOrigin);
-        res.status(200).json({ greeting });
-        return;
-      } catch (greetingErr) {
-        logError("Failed to generate proactive greeting", greetingErr);
-        res.setHeader("Access-Control-Allow-Origin", validOrigin);
-        res.status(200).json({ greeting: null, reason: "generation failed" });
-        return;
-      }
+      res.setHeader("Access-Control-Allow-Origin", validOrigin);
+      res.status(proactiveResult.status).json(proactiveResult.body);
+      return;
     }
 
     // -------------------------------------------------------------
